progress: 21.94% of code matched (3190/8216 functions, 3424/4775 objects)

+++ src/game/game_fn_80172610.c
@@ -1,5 +1,5 @@
 extern int fn_8016A598(void*);
-extern float fn_8016A694(void*, int);
+extern double fn_8016A694(void*, int);
 extern void fn_80163BB4(void*, const char*, ...);
 extern void fn_801AC350(signed char, int, int);
 extern const char lbl_8024FF00[];
