+++ src/game/game_fn_801726EC.c
@@ -11,7 +11,7 @@ typedef struct Vec3f {
 } Vec3f;
 
 extern int fn_8016A598(void*);
-extern float fn_8016A694(void*, int);
+extern double fn_8016A694(void*, int);
 extern Vec3s* fn_8016A784(void*, int);
 extern void fn_8016A830(void*, double);
 extern void fn_80163BB4(void*, const char*, ...);
